# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: SeatShare ride-sharing platform API
servers:
  - url: /api
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      summary: Send OTP to phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      summary: Verify OTP and get tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
...
  /driver-profiles:
    get:
      operationId: listDriverProfiles
      summary: List driver profiles
      parameters:
        - in: query
          name: isOnline
          schema:
            type: boolean
        - in: query
          name: isVerified
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of driver profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfileList"
  /driver-profiles/{id}:
    get:
      operationId: getDriverProfile
      summary: Get driver profile by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
    patch:
      operationId: updateDriverProfile
      summary: Update driver profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /driver-profiles/{id}/verify:
    patch:
      operationId: verifyDriverProfile
      summary: Verify or unverify a driver
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
...
  /bookings:
    get:
      operationId: listBookings
      summary: List bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tripId
          schema:
            type: integer
        - in: query
          name: passengerId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
    post:
      operationId: createBooking
      summary: Passenger requests a seat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      summary: Get booking by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/accept:
    patch:
      operationId: acceptBooking
      summary: Driver accepts a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInput"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/reject:
    patch:
      operationId: rejectBooking
      summary: Driver rejects a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/complete:
    patch:
      operationId: completeBooking
      summary: Mark booking as completed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteInput"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /payments:
    get:
      operationId: listPayments
      summary: List payments
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: bookingId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
    post:
      operationId: createPayment
      summary: Initiate payment for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      summary: Get payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpResponse:
      type: object
      required: [message, sessionId]
      properties:
        message:
          type: string
        sessionId:
          type: string
    OtpVerify:
      type: object
      required: [phone, otp, sessionId]
      properties:
        phone:
          type: string
        otp:
          type: string
        sessionId:
          type: string
        role:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
...
    DriverProfile:
      type: object
      required: [id, userId, isVerified, isOnline, rating, totalTrips, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        licenseNumber:
          type: ["string", "null"]
        rating:
          type: number
        totalTrips:
          type: integer
        isVerified:
          type: boolean
        isOnline:
          type: boolean
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        user:
          $ref: "#/components/schemas/User"
        vehicle:
          oneOf:
            - $ref: "#/components/schemas/Vehicle"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    DriverProfileList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DriverProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DriverProfileUpdate:
      type: object
      properties:
        licenseNumber:
          type: string
        currentLat:
          type: number
        currentLng:
          type: number
    VerifyInput:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
    Vehicle:
      type: object
      required: [id, driverProfileId, make, model, year, color, licensePlate, capacity, createdAt]
      properties:
        id:
          type: integer
        driverProfileId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        capacity:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Trip:
      type: object
      required: [id, driverProfileId, originAddress, destAddress, originLat, originLng, destLat, destLng, availableSeats, maxDeviationKm, status, farePerSeat, departureTime, createdAt]
      properties:
        id:
          type: integer
        driverProfileId:
          type: integer
        originAddress:
          type: string
        destAddress:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
        availableSeats:
          type: integer
        maxDeviationKm:
          type: number
        status:
          type: string
        farePerSeat:
          type: number
        departureTime:
          type: string
          format: date-time
        driverProfile:
          oneOf:
            - $ref: "#/components/schemas/DriverProfile"
            - type: "null"
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      required: [id, tripId, passengerId, pickupAddress, dropoffAddress, pickupLat, pickupLng, dropoffLat, dropoffLng, status, fare, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        passengerId:
          type: integer
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        dropoffLat:
          type: number
        dropoffLng:
          type: number
        status:
          type: string
        fare:
          type: number
        boardingCode:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        trip:
          oneOf:
            - $ref: "#/components/schemas/Trip"
            - type: "null"
        passenger:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    BookingList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BookingInput:
      type: object
      required: [tripId, passengerId, pickupAddress, dropoffAddress, pickupLat, pickupLng, dropoffLat, dropoffLng]
      properties:
        tripId:
          type: integer
        passengerId:
          type: integer
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        dropoffLat:
          type: number
        dropoffLng:
          type: number
    AcceptInput:
      type: object
      properties:
        note:
          type: string
    RejectInput:
      type: object
      properties:
        reason:
          type: string
    CompleteInput:
      type: object
      properties:
        boardingCode:
          type: string
    Payment:
      type: object
      required: [id, bookingId, amount, status, method, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        amount:
          type: number
        status:
          type: string
        method:
          type: string
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PaymentList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PaymentInput:
      type: object
      required: [bookingId, amount, method]
      properties:
        bookingId:
          type: integer
        amount:
          type: number
        method:
          type: string
...