# lib/api-zod/src/generated/types/driverProfileList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeatShare ride-sharing platform API
 * OpenAPI spec version: 1.0.0
 */
import type { DriverProfile } from './driverProfile';

export interface DriverProfileList {
  data: DriverProfile[];
  total: number;
  page: number;
  limit: number;
}