# lib/api-zod/src/generated/types/paymentList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SeatShare ride-sharing platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Payment } from './payment';

export interface PaymentList {
  data: Payment[];
  total: number;
  page: number;
  limit: number;
}